docs/trading_ai_system_updated/trading_ai_system/control_service/requirements.txt
kafka-python>=2.0.0
ccxt>=4.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
tenacity>=8.2.3